Fix sign-compare warning in nwrite.h
This fixes the following gcc 8.1.0 compilier warning with -Wextra:

../lib/nwrite.h: In function ‘nwrite’:
../lib/nwrite.h:34:13: warning: comparison of integer expressions of different signedness: ‘ssize_t’ {aka ‘long int’} and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
  while (tot < count) {
             ^

+++ lib/nwrite.h
@@ -26,12 +26,17 @@
  */
 static inline ssize_t nwrite(int fd, const void *buf, size_t count, ssize_t *written)
 {
+	/*
+	 * Given the API we have to assume (unsigned) size_t 'count' fits into
+	 * a (signed) ssize_t because we can't return a larger value.
+	 * https://stackoverflow.com/questions/29722999/will-write2-always-write-less-than-or-equal-to-ssize-max
+	 */
 	ssize_t	out, tot = 0;
 
 	if (!buf || count == 0)
 		return 0;
 
-	while (tot < count) {
+	while ((size_t) tot < count) {
 		out = write(fd, (const char *) buf + tot, count - tot);
 		if (out > 0)
 			tot += out;
